move setting to privacymanager
